+++ packages/core/src/components/viewer-grid/components/document-panel.tsx
@@ -1,10 +1,7 @@
 import { EventEmitter, FunctionalComponent, h } from "@stencil/core";
 
-import {
-  ViewerGridPanelSize,
-  ViewerGridChangeSizeAnimationEndEvent,
-  viewerGridSizeLabelMap,
-} from "../viewer-grid.interfaces";
+import { ViewerGridPanelSize, ViewerGridChangeSizeAnimationEndEvent } from "../viewer-grid.interfaces";
+import { SizingButtons } from "./sizing-buttons";
 
 export interface ViewerGridDocumentPanelProps {
   tabView: boolean;
@@ -30,23 +27,12 @@ export const DocumentPanel: FunctionalComponent<ViewerGridDocumentPanelProps> = 
     }}
   >
     {!tabView && (
-      <div class="sizing-buttons">
-        <span class="sr-only" aria-live="polite" aria-atomic="true">
-          Breedte documentpaneel: {viewerGridSizeLabelMap[panelSize]}
-        </span>
-        {panelSize !== "small" && (
-          <button type="button" class="shrink" onClick={shrinkDocumentPanel}>
-            <span class="sr-only">Documentpaneel smaller maken</span>
-            <dso-icon icon="chevron-right"></dso-icon>
-          </button>
-        )}
-        {panelSize !== "large" && (
-          <button type="button" class="expand" onClick={expandDocumentPanel}>
-            <span class="sr-only">Documentpaneel breder maken</span>
-            <dso-icon icon="chevron-left"></dso-icon>
-          </button>
-        )}
-      </div>
+      <SizingButtons
+        panelLabel="Documentpaneel"
+        size={panelSize}
+        expand={expandDocumentPanel}
+        shrink={shrinkDocumentPanel}
+      />
     )}
     <div class="content">
       <slot name="document-panel" />