+++ packages/core/src/components/viewer-grid/components/main-panel.tsx
@@ -1,12 +1,8 @@
 import { EventEmitter, FunctionalComponent, h } from "@stencil/core";
 import clsx from "clsx";
 
-import {
-  ViewerGridPanelSize,
-  ViewerGridChangeSizeAnimationEndEvent,
-  ViewerGridMode,
-  viewerGridSizeLabelMap,
-} from "../viewer-grid.interfaces";
+import { ViewerGridPanelSize, ViewerGridChangeSizeAnimationEndEvent, ViewerGridMode } from "../viewer-grid.interfaces";
+import { SizingButtons } from "./sizing-buttons";
 
 export interface ViewerGridMainPanelProps {
   mode: ViewerGridMode;
@@ -48,34 +44,24 @@ export const MainPanel: FunctionalComponent<ViewerGridMainPanelProps> = ({
     }}
   >
     {!tabView &&
-      (!documentPanelOpen ? (
-        <div class="sizing-buttons">
-          <span class="sr-only" aria-live="polite" aria-atomic="true">
-            Breedte {mode === "vdk" ? "zoeken paneel" : "hoofdpaneel"}: {viewerGridSizeLabelMap[mainSize]}
-          </span>
-          {mainSize !== "small" && (
-            <button type="button" class="shrink" onClick={shrinkMain}>
-              <span class="sr-only">{mode === "vdk" ? "Zoeken paneel" : "Hoofdpaneel"} smaller maken</span>
-              <dso-icon icon="chevron-left"></dso-icon>
+      (((mode === "vrk" || documentPanelOpen) && (
+        <SizingButtons
+          panelLabel={mode === "vdk" ? "Zoeken paneel" : "Hoofdpaneel"}
+          size={mainSize}
+          expand={expandMain}
+          shrink={shrinkMain}
+        />
+      )) ||
+        (mode === "vdk" && documentPanelOpen && (
+          <div class="toggle-button">
+            <button type="button" onClick={toggleMainPanel}>
+              <span class="sr-only">
+                {mode === "vdk" ? "Zoeken paneel" : "Hoofdpaneel"} {mainPanelHidden ? "Tonen" : "Verbergen"}
+              </span>
+              <dso-icon icon={mainPanelHidden ? "chevron-right" : "chevron-left"}></dso-icon>
             </button>
-          )}
-          {mainSize !== "large" && (
-            <button type="button" class="expand" onClick={expandMain}>
-              <span class="sr-only">{mode === "vdk" ? "Zoeken paneel" : "Hoofdpaneel"} breder maken</span>
-              <dso-icon icon="chevron-right"></dso-icon>
-            </button>
-          )}
-        </div>
-      ) : (
-        <div class="toggle-button">
-          <button type="button" onClick={toggleMainPanel}>
-            <span class="sr-only">
-              {mode === "vdk" ? "Zoeken paneel" : "Hoofdpaneel"} {mainPanelHidden ? "Tonen" : "Verbergen"}
-            </span>
-            <dso-icon icon={mainPanelHidden ? "chevron-right" : "chevron-left"}></dso-icon>
-          </button>
-        </div>
-      ))}
+          </div>
+        )))}
     <div class={clsx("content", { invisible: mainPanelHidden })}>
       <slot name="main" />
       {!tabView && documentPanelOpen && (