+++ CHANGELOG.md
@@ -27,6 +27,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Fixed
 * Date Picker: Input en output worden niet prepadded met "0" ([#2409](https://github.com/dso-toolkit/dso-toolkit/issues/2409))
+* Viewer Grid (VDK): Diverse bevindingen ([#2410](https://github.com/dso-toolkit/dso-toolkit/issues/2410))
 
 ## 🎤 62.0.0 - 20-10-2023