+++ CHANGELOG.md
@@ -29,6 +29,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Fixed
 * Date Picker: Input en output worden niet prepadded met "0" ([#2409](https://github.com/dso-toolkit/dso-toolkit/issues/2409))
 
+### Fixed
+* Image: Large image overflows container ([#2401](https://github.com/dso-toolkit/dso-toolkit/issues/2401))
+
 ## 🎤 62.0.0 - 20-10-2023
 
 ## Added