+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Fixed
 * Anchor: Icon no wrap ([#2329](https://github.com/dso-toolkit/dso-toolkit/issues/2329))
 * Viewer Grid (VDK): Diverse bevindingen ([#2410](https://github.com/dso-toolkit/dso-toolkit/issues/2410))
+* Image: Large image overflows container ([#2401](https://github.com/dso-toolkit/dso-toolkit/issues/2401))
 
 ### Added
 * Header: "Help"-button toevoegen ([#2334](https://github.com/dso-toolkit/dso-toolkit/issues/2334))
@@ -28,9 +29,6 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Fixed
 * Date Picker: Input en output worden niet prepadded met "0" ([#2409](https://github.com/dso-toolkit/dso-toolkit/issues/2409))
-
-### Fixed
-* Image: Large image overflows container ([#2401](https://github.com/dso-toolkit/dso-toolkit/issues/2401))
 
 ## 🎤 62.0.0 - 20-10-2023